# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EGH Panel API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Nests
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, client]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - username
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            servers:
              type: array
              items:
                $ref: "#/components/schemas/Server"
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, client]
      required:
        - email
        - username
        - firstName
        - lastName
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, client]
        isActive:
          type: boolean
        password:
          type: string
...
    Server:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        userId:
          type: integer
        userEmail:
          type: string
        nodeId:
          type: integer
        nodeName:
          type: string
        eggId:
          type: integer
        eggName:
          type: string
        allocationId:
          type: integer
        allocationIp:
          type: string
        allocationPort:
          type: integer
        status:
          type: string
          enum: [installing, install_failed, running, offline, stopping, starting, suspended]
        memoryLimit:
          type: integer
        diskLimit:
          type: integer
        cpuLimit:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - uuid
        - name
        - userId
        - userEmail
        - nodeId
        - nodeName
        - eggId
        - eggName
        - allocationId
        - allocationIp
        - allocationPort
        - status
        - memoryLimit
        - diskLimit
        - cpuLimit
        - createdAt
...